Refactor groups folder structure
