+++ src/README.md
@@ -4,3 +4,8 @@
 - `yarn` to install
 - `yarn dev` To run local server
 - `yarn build` To make a build
+
+## Default Super Admin
+
+DNI:  74237694L
+Password: academia750
