+++ src/components/navigation/MainMenu.vue
@@ -2,13 +2,12 @@
   <div v-if="user">
     <v-divider></v-divider>
     <v-list-item class="py-1 px-3" two-line>
-      <v-list-item-avatar
-        class="ml-n1"
-        min-width="40"
-        height="40"
-        width="40"
-      >
-        <v-img v-if="pathImageAccount" :src="pathImageAccount" @error="loadImageProfileDefault"/>
+      <v-list-item-avatar class="ml-n1" min-width="40" height="40" width="40">
+        <v-img
+          v-if="pathImageAccount"
+          :src="pathImageAccount"
+          @error="loadImageProfileDefault"
+        />
         <v-skeleton-loader
           v-else-if="!pathImageAccount && !hasErrorImage"
           class="mx-auto"
@@ -18,9 +17,7 @@
       </v-list-item-avatar>
 
       <v-list-item-content>
-        <v-list-item-title
-          class="text-16 font-weight-regular"
-        >
+        <v-list-item-title class="text-16 font-weight-regular">
           {{ user.attributes.first_name }}
         </v-list-item-title>
         <!-- <v-divider class="d-block" /> -->
@@ -29,7 +26,14 @@
     <v-divider></v-divider>
     <v-list v-if="user" nav dense>
       <div v-for="(item, index) in menu" :key="index">
-        <div v-if="(item.key || item.text) && ( $can(item.can) || $hasRoles(item.roles) )" class="pa-1 mt-2 overline">{{ item.key ? $t(item.key) : item.text }}</div>
+        <div
+          v-if="
+            (item.key || item.text) && ($can(item.can) || $hasRoles(item.roles))
+          "
+          class="pa-1 mt-2 overline"
+        >
+          {{ item.key ? $t(item.key) : item.text }}
+        </div>
         <nav-menu :menu="item.items" />
       </div>
     </v-list>
@@ -50,7 +54,7 @@ export default {
       default: () => []
     }
   },
-  data () {
+  data() {
     return {
       pathImageAccount: null,
       hasErrorImage: false
@@ -60,7 +64,7 @@ export default {
     ...mapState('profileService', ['user'])
   },
   watch: {
-    user (value) {
+    user(value) {
       if (!value) {
         return
       }
@@ -68,18 +72,18 @@ export default {
       this.loadImageAccount()
     }
   },
-  created () {
+  created() {
     this.loadImageAccount()
     /* //console.log(this.$userAuth())
     //console.log(this.$rolesUserAuth())
     //console.log(this.$permissionsUserAuth()) */
   },
   methods: {
-    loadImageProfileDefault () {
+    loadImageProfileDefault() {
       this.hasErrorImage = true
       this.pathImageAccount = '/images/academia750/avatar_default_photo.svg'
     },
-    loadImageAccount () {
+    loadImageAccount() {
       const IsDevelopmentEnviroment = process.env.NODE_ENV === 'development'
       const serverApiDevelopment = process.env.VUE_APP_BASE_URL_API_DEVELOPMENT
       const serverApiProduction = process.env.VUE_APP_BASE_URL_API_PRODUCTION
@@ -98,7 +102,9 @@ export default {
       }
 
       if (image.attributes.type_path === 'local') {
-        this.pathImageAccount = `${IsDevelopmentEnviroment ? serverApiDevelopment : serverApiProduction}${image.attributes.path}`
+        this.pathImageAccount = `${
+          IsDevelopmentEnviroment ? serverApiDevelopment : serverApiProduction
+        }${image.attributes.path}`
 
         return
       }
