--- src/modules/groups/components/datatable/component/ActionsMethods.js
@@ -1,30 +1,0 @@
-export default {
-  methods: {
-    async deleteStudentAction () {
-      if (this.currentItemsSelectedForDelete === null || this.currentItemsSelectedForDelete === undefined) {
-        return
-      }
-
-      try {
-        await this.deleteStudent({
-          id: this.currentItemsSelectedForDelete.id,
-          config: {}
-        })
-
-        this.$swal.fire({
-          icon: 'success',
-          toast: true,
-          title: 'El alumno ha sido eliminado con éxito.',
-          timer: 3000
-        })
-
-        this.$refs['ResourceTextFieldSearch'].searchWordText = ''
-        this.searchWord = ''
-
-        this.loadStudentsFromCurrentTab()
-      } catch (error) {
-        //console.log(error)
-      }
-    }
-  }
-}
