--- src/modules/groups/components/datatable/component/data.js
@@ -1,8 +1,0 @@
-export default {
-  data () {
-    return {
-      //namesRelationshipsIncludeRequest: 'topics'
-      currentItemsSelectedForDelete: null
-    }
-  }
-}
