--- src/modules/groups/components/datatable/component/props.js
@@ -1,8 +1,0 @@
-export default {
-  props: {
-    stateAccount: {
-      type: String,
-      default: 'enable'
-    }
-  }
-}
