--- src/modules/groups/components/form/FormManageStatusAccount.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>
-
-  </div>
-</template>
-
-<script>
-export default {
-
-}
-</script>
