+++ src/modules/groups/index.js
@@ -1,4 +1,4 @@
-import store from './store/index'
+import store from './store'
 import router from './router'
 
 export default {
