--- src/modules/groups/store/actions.js
@@ -1,168 +1,0 @@
-import GroupRepository from '../repositories/GroupRepository'
-
-const mapItemsDatatableFromApi = (itemsApi) => {
-  return itemsApi.map((record) => {
-    return {
-      id: record.id,
-      dni: record.attributes.dni,
-      'first-name': record.attributes.first_name,
-      'last-name': record.attributes.last_name,
-      phone: record.attributes.phone,
-      email: record.attributes.email,
-      'created-at': record.attributes.created_at
-    }
-  })
-}
-
-const mapMetaInformationPagination = (response) => {
-  return {
-    current_page: response.data.meta.current_page,
-    from: response.data.meta.from,
-    last_page: response.data.meta.last_page,
-    per_page: response.data.meta.per_page,
-    to: response.data.meta.to,
-    total: response.data.meta.total
-  }
-}
-
-const getStudents = async ({ commit }, config) => {
-  try {
-    commit('SET_USERS_SELECTED_DATATABLE', [])
-    commit('SET_INFORMATION_META', {
-      current_page: 1,
-      from: 1,
-      last_page: 1,
-      per_page: 10,
-      to: 10,
-      total: 10
-    })
-    commit('SET_STATUS_LOADING_ITEMS', true)
-
-    commit('SET_ITEMS_DATATABLE', [])
-
-    const response = await GroupRepository.getAll(config)
-
-    if (response) {
-      //console.trace(response)
-
-      commit(
-        'SET_ITEMS_DATATABLE',
-        mapItemsDatatableFromApi(response.data.data)
-      )
-      commit('SET_STATUS_LOADING_ITEMS', false)
-      commit('SET_INFORMATION_META', mapMetaInformationPagination(response))
-    }
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const createStudent = async (_, options) => {
-  try {
-    const response = await GroupRepository.create(options.data)
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const fetchStudent = async ({ commit }, options) => {
-  try {
-    commit('SET_USERS_SELECTED_DATATABLE', [])
-    const response = await GroupRepository.get(options.id, options.config)
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const updateStudent = async (_, options) => {
-  try {
-    const response = await GroupRepository.update(
-      options.id,
-      options.data,
-      options.config
-    )
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const deleteStudent = async (_, options) => {
-  try {
-    const response = await GroupRepository.delete(options.id, options.config)
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const fetchStudentGroups = async (_, options) => {
-  try {
-    const response = await GroupRepository.fetchStudentsGroups(options.config)
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const fetchRoleStudentData = async (_, options) => {
-  try {
-    const response = await GroupRepository.fetchRoleStudentData(options.config)
-
-    return Promise.resolve(response)
-  } catch (error) {
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-const actionsForMultipleRecords = async ({ commit }, options) => {
-  try {
-    const response = await GroupRepository.actionMassiveSelection(
-      options.data,
-      options.config
-    )
-
-    commit('SET_USERS_SELECTED_DATATABLE', [])
-
-    return Promise.resolve(response)
-  } catch (error) {
-    commit('SET_USERS_SELECTED_DATATABLE', [])
-    //console.log(error)
-
-    return Promise.reject(error)
-  }
-}
-
-export default {
-  getStudents,
-  createStudent,
-  fetchStudent,
-  updateStudent,
-  deleteStudent,
-  fetchStudentGroups,
-  fetchRoleStudentData,
-  actionsForMultipleRecords
-}
