--- src/modules/groups/store/getters.js
@@ -1,2 +1,0 @@
-export default {
-}
