--- src/modules/groups/store/index.js
@@ -1,12 +1,0 @@
-import state from './state'
-import mutations from './mutations'
-import getters from './getters'
-import actions from './actions'
-
-export default {
-  namespaced: true,
-  state,
-  getters,
-  mutations,
-  actions
-}
